pyproject.toml
[project]
name = "biz2bricks-core"
version = "0.1.0"
description = "Shared core library for Biz2Bricks: models, database connection, and migrations"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Biz2Bricks Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Database (PostgreSQL/Cloud SQL)
    "sqlalchemy[asyncio]==2.0.44",
    "asyncpg==0.31.0",
    "cloud-sql-python-connector[asyncpg]==1.18.5",
    # Migrations
    "alembic>=1.13.0",
    # Configuration
    "pydantic>=2.12.5,<3.0.0",
    "pydantic-settings>=2.7.1",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/biz2bricks_core"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/migrations",
    "/alembic.ini",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
